refactor(tools): extract package and unify agent tool execution
Move tools.py into tools/ package (base, bash, file modules).
Encapsulate bash streaming in bash_handler so agent.py follows a single unified execution path for all tools.
