+++ src/mini_agent/agent/tools/file.py
@@ -1,0 +1,35 @@
+from .base import MAX_OUTPUT, resolve_path
+
+
+def run_read(path: str, limit: int | None = None) -> str:
+    try:
+        text = resolve_path(path).read_text()
+        lines = text.splitlines()
+        if limit and limit < len(lines):
+            remaining = len(lines) - limit
+            lines = lines[:limit] + [f"... ({remaining} more lines)"]
+        return "\n".join(lines)[:MAX_OUTPUT]
+    except Exception as exc:
+        return f"Error: {exc}"
+
+
+def run_write(path: str, content: str) -> str:
+    try:
+        file_path = resolve_path(path)
+        file_path.parent.mkdir(parents=True, exist_ok=True)
+        file_path.write_text(content)
+        return f"Wrote {len(content)} bytes to {path}"
+    except Exception as exc:
+        return f"Error: {exc}"
+
+
+def run_edit(path: str, old_text: str, new_text: str) -> str:
+    try:
+        file_path = resolve_path(path)
+        content = file_path.read_text()
+        if old_text not in content:
+            return f"Error: Text not found in {path}"
+        file_path.write_text(content.replace(old_text, new_text, 1))
+        return f"Edited {path}"
+    except Exception as exc:
+        return f"Error: {exc}"
