+++ src/mini_agent/cli/display/printing.py
@@ -106,10 +106,7 @@ def print_tool_start(name: str, input_data: dict[str, object]) -> None:
 def print_tool_result(name: str, input_data: dict[str, object], output: str) -> None:
     """Print tool output after execution."""
     if name == "bash":
-        text = Text.from_ansi(output)
-        text.stylize(LIGHT_HINT_STYLE_RICH)
-        console.print(text)
-        print()
+        console.print()
         return
 
     if name == "edit_file":
